restructure to avoid circular import
